main_test.go
package main

import (
	"net/http"
	"net/http/httptest"
	"testing"
)

func TestMyHandler(t *testing.T) {
	// Create a request to test the handler
	req := httptest.NewRequest(http.MethodGet, "/my-endpoint", nil)
	// Create a response recorder to capture the response
	rr := httptest.NewRecorder()

	// Call the handler function
	MyHandler(rr, req)

	// Assert the response status code is 204
	if rr.Code != http.StatusNoContent {
		t.Errorf("expected status %d, got %d", http.StatusNoContent, rr.Code)
	}
}

func TestMyConcurrentFunction(t *testing.T) {
	// Create a channel to signal when the concurrent function is done
	done := make(chan bool)

	// Execute the concurrent function
	go func() {
		MyConcurrentFunction()
		done <- true
	}()

	// Wait for the concurrent function to complete
	<-done

	// Add assertions or verifications here
}

func TestHandlerIntegration(t *testing.T) {
	// Create a new HTTP test server with the handler
	srv := httptest.NewServer(http.HandlerFunc(MyHandler))
	defer srv.Close()

	// Make an HTTP request to the app
	resp, err := http.Get(srv.URL + "/my-endpoint")
	if err != nil {
		t.Fatalf("failed to make request: %v", err)
	}
	defer resp.Body.Close()

	// Assert the response status code is 204
	if resp.StatusCode != http.StatusNoContent {
		t.Errorf("expected status %d, got %d", http.StatusNoContent, resp.StatusCode)
	}
}
